% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_desc.R
\name{fill_desc}
\alias{fill_desc}
\title{Fill your description}
\usage{
fill_desc(
  pkg_title,
  pkg_description,
  author_first_name,
  author_last_name,
  author_email,
  repo_url = NULL,
  pkg = here::here()
)
}
\arguments{
\item{pkg_title}{Full title of the package}

\item{pkg_description}{Description of the package}

\item{author_first_name}{First Name of the author}

\item{author_last_name}{Last Name of the author}

\item{author_email}{Email of the author}

\item{repo_url}{URL (if needed)}
}
\description{
Fill your description
}
