% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_desc.R
\name{if_not_null}
\alias{if_not_null}
\title{Check for NULL values}
\usage{
if_not_null(x, ...)
}
\arguments{
\item{x}{Value to check.}

\item{...}{Other arguments to be passed.}
}
\description{
Check for NULL values
}
