% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_d6_packages.R
\name{install_d6_packages}
\alias{install_d6_packages}
\title{Install common packages used in the D6 department}
\usage{
install_d6_packages(geo = TRUE)
}
\arguments{
\item{geo}{Boolean. Do you intent to process geodata? Installs important geo packages if `TRUE` (the default)}
}
\description{
Install common packages used in the D6 department
}
