% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{new_project}
\alias{new_project}
\title{Create new project scaffolding.}
\usage{
new_project(name, path = getwd(), github = FALSE, private.repo = TRUE)
}
\arguments{
\item{name}{Name of the new project. A new folder will be created with that name.}

\item{path}{path of the new project where the folder is created. If empty then the current working directory is used.}

\item{github}{Logical. Create GitHub repo? Note this requires a \code{GITHUB_PAT}. See \code{\link[devtools]{use_github}}.}

\item{private.repo}{Logical. Default is TRUE.}
}
\value{
A new directory with R package structure, slightly modified.
}
\description{
Create all the scaffolding for a new project in a new directory. The scaffolding includes a \code{README.Rmd} file, different folders to hold raw data, analyses, etc, and \code{testthat} infrastructure. Also, optionally, set a GitHub repo and add Travis-CI.
}
