% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_all_reports}
\alias{render_all_reports}
\title{Render All Rmd files placed in ./R to html in ./docs/reports}
\usage{
render_all_reports(sub_dirs = FALSE)
}
\arguments{
\item{sub_dirs}{Should Rmd files placed in sub directories (./R/.../*.Rmd) be
rendered? Defaults to FALSE.}
}
\description{
Render All Rmd files placed in ./R to html in ./docs/reports
}
