% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_load.R
\name{simple_load}
\alias{simple_load}
\title{Load Multiple Packages in One Go (and install them if necessary)}
\usage{
simple_load(pckgs = NULL, update_cran = FALSE, update_gh = FALSE)
}
\arguments{
\item{pckgs}{A vector of package names. Packages from CRAN and GitHub can 
be passed simultaneously. For GitHub packages, pass the package formatted as 
"owner/repository".}

\item{update_cran}{Logical. Should CRAN packages be updated? Defaults to FALSE.}

\item{update_gh}{Logical. Should GitHub packages be updated? Defaults to FALSE.}
}
\description{
Load Multiple Packages in One Go (and install them if necessary)
}
\examples{
simple_load("ggplot2", "EcoDynIZW/d6berlin")

}
