% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_d6.R
\name{theme_d6}
\alias{theme_d6}
\title{A custom ggplot2 theme.}
\usage{
theme_d6(
  base_size = 14,
  base_family = "PT Sans",
  grid = "xy",
  legend = "bottom",
  mono = "none",
  bg = "transparent",
  serif = FALSE,
  margin = rep(base_size/2, 4),
  ...
)
}
\arguments{
\item{base_size}{Numeric. The base sizes used for theme elements, defaults to 14 
(which is slightly larger than in the default themes which use 11).}

\item{base_family}{Character. The base font family, defaults to PT Sans. 
Typefaces need to be installed locally.}

\item{grid}{Character. Should the panel grid be drawn? Specify \code{""} or 
code{"none"} 
to draw no grid or any combination of \code{"x"}, \code{"X"}, \code{"y"}, 
and \code{"Y"} to draw horizontal (y) and/or vertical (x) grid lines.}

\item{legend}{Character or numeric vector. The position of legends as to be 
used in the \code{`theme()`} function: either \code{"none"}, \code{"left"}, 
\code{"right"}, \code{"bottom"}, or \code{"top"}, or a two-element numeric 
vector specifying the relative x and y coordinates to place the legend inside 
the panel.}

\item{mono}{Character. Should the axis and/or legend text be formatted as tabular, 
mono-spaced font? Specify the axes with \code{"x"}, \code{"X"}, \code{"y"}, 
and \code{"Y"} and the legend as \code{"legend"}, \code{"l"} or \code{"L"}. 
If you want to use the base family, set \code{""} or \code{"none"} (default).}

\item{bg}{Color. The color used for the plot background, defaults to 
transparent.}

\item{serif}{Logical. Should Sans or Sans Serif fonts be used? Default is 
\code{FALSE}.}

\item{margin}{Vector. The plot margin four-element numeric vector specifying 
the margin on the top, right, bottom, and left. Defaults to 
\code{rep(base_size / 2, 4)}}

\item{...}{Other arguments passed to complete ggplot2 themes.}
}
\description{
A dedicated theme for graphics created within the D6, with sensible defaults 
and additional options. It is based on the light theme and uses the custom 
typefaces PT Sans (defaul) and PT Serif (optional).
}
\examples{
library(ggplot2)
ggplot(mpg, aes(x = class, y = cty, color = class)) + geom_boxplot() +
  theme_d6(grid = "y", bg = "grey90", legend = "top", mono = "y")
}
