% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_report.R
\name{render_report}
\alias{render_report}
\title{Render Rmd placed in ./R to html in ./docs/reports}
\usage{
render_report(rmd)
}
\arguments{
\item{rmd}{A string. Name of the Rmd file}
}
\description{
Render Rmd placed in ./R to html in ./docs/reports
}
