% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_load.R
\name{simple_load}
\alias{simple_load}
\title{Load Multiple Packages in One Go (and install them if necessary)}
\usage{
simple_load(pckgs = NULL)
}
\arguments{
\item{pckgs}{A vector of packages to load. Packages from CRAN and GitHub can 
be passed simultaneously.}
}
\description{
Load Multiple Packages in One Go (and install them if necessary)
}
\examples{
simple_load("ggplot2")

}
